# docs/youapp-swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/register": {
      "post": {
        "operationId": "AppController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "operationId": "AppController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/getProfile": {
      "get": {
        "operationId": "AppController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/updateProfile": {
      "put": {
        "operationId": "AppController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/askHoroscopeZodiac": {
      "post": {
        "operationId": "AppController_askHoroscopeZodiac",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/profiles": {
      "get": {
        "operationId": "AppController_getAllProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/refresh": {
      "get": {
        "operationId": "AppController_refreshToken",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    }
  },
  "info": {
    "title": "Microservice name",
    "description": "The microservice API description",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "The username of the new user",
            "example": "johndoe"
          },
          "email": {
            "type": "string",
            "description": "The email of the new user",
            "example": "johndoe@example.com"
          },
          "password": {
            "type": "string",
            "description": "The password of new the user",
            "example": "123456"
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ]
      },
      "LoginUserDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}